+++ core_gearworks_lib/source/lib/presets/2dim/default/p2d_def.h
@@ -8,12 +8,6 @@
 
 class defshape {
 private:
-	// Vectors to cache transformations that have not yet been applied.
-	// These get released in the private apply_transformations() function, which is run directly before glDrawElements().
-	glm::vec3 active_translation;
-	glm::vec3 active_scaling;
-	glm::vec3 active_rotation;
-
 	// This function applies any outstanding transformations to the shape.
 	// It is only ever run in the base render() function, and so it can be private.
 	// In order to publicly transform the primitive, use the translate(), scale(), and rotate() functions.
@@ -27,21 +21,45 @@ protected:
 	gearworks::index_buffer_object ibo;
 
 	// Primitive type:- (for some functions in render())
-	//	 0 = triangle
-	//	 1 = rectangle
+	// | 0 = triangle
+	// | 1 = rectangle
 	unsigned int primitive_type;
 
 	// Texture used for the primitive
 	gearworks::texture_2d *texture_object;
 public:
 	/// <summary>
-	/// The blending colour of the shape
+	/// A struct that holds properties of the shape. Don't reference this in your own program; use the shape::properties object instead.
 	/// </summary>
-	glm::vec4 colour;
+	struct shape_properties_struct {
+	public:
+		// Constructor for a shape_properties_struct object.
+		shape_properties_struct();
+
+		/// <summary>
+		/// The blending colour of the shape.
+		/// </summary>
+		vec3 colour;
+		/// <summary>
+		/// The opacity of the shape.
+		/// </summary>
+		float opacity;
+
+		/// <summary>
+		/// The path to the texture of the shape.
+		/// </summary>
+		string texture;
+
+		/// <summary>
+		/// Basic transformation value of the shape.
+		/// </summary>
+		vec3 position, scale, rotation;
+	};
+
 	/// <summary>
-	/// The path to the texture of the shape
+	/// The properties of the shape, like colour, texture, etc.
 	/// </summary>
-	std::string texture;
+	shape_properties_struct properties;
 
 	/// <summary>
 	/// Default constructor. Note: to actually create the rectangle with vertex coordinates etc, use the create() function.
@@ -62,12 +80,12 @@ public:
 	/// <param name="x">Value to move on the X axis.</param>
 	/// <param name="y">Value to move on the Y axis.</param>
 	/// <param name="z">Value to move on the Z axis.</param>
-	inline void translate(float x, float y, float z) { active_translation += glm::vec3(x, y, z); }
+	inline void translate(float x, float y, float z) { properties.position += vec3(x, y, z); }
 	/// <summary>
 	/// Translates, or moves, the shape by a given distance.
 	/// </summary>
 	/// <param name="vect">Value to move on the 3 axes stored as a glm::vec3.</param>
-	inline void translate(glm::vec3 vect) { active_translation += vect; }
+	inline void translate(vec3 vect) { properties.position += vect; }
 
 	/// <summary>
 	/// Rotates the shape by a given amount.
@@ -75,12 +93,12 @@ public:
 	/// <param name="x">Value to rotate on the X axis.</param>
 	/// <param name="y">Value to rotate on the Y axis.</param>
 	/// <param name="z">Value to rotate on the Z axis.</param>
-	inline void rotate(float x, float y, float z) { active_rotation -= glm::radians(glm::vec3(x, y, z)); } // subtracted to rotate clockwise by default
+	inline void rotate(float x, float y, float z) { properties.rotation += vec3(x, y, z); } // subtracted to rotate clockwise by default
 	/// <summary>
 	/// Rotates the shape by a given amount.
 	/// </summary>
 	/// <param name="vect">Value to rotate the shape on the 3 axes stored as a glm::vec3.</param>
-	inline void rotate(glm::vec3 vect) { active_rotation -= glm::radians(vect); } // subtracted to rotate clockwise by default
+	inline void rotate(vec3 vect) { properties.rotation += vect; } // subtracted to rotate clockwise by default
 
 	/// <summary>
 	/// Scales the shape by a given amount.
@@ -88,17 +106,17 @@ public:
 	/// <param name="x">Value to scale on the X axis.</param>
 	/// <param name="y">Value to scale on the Y axis.</param>
 	/// <param name="z">Value to scale on the Z axis.</param>
-	inline void scale(float x, float y, float z) { active_scaling += glm::vec3(x, y, z); }
+	inline void scale(float x, float y, float z) { properties.scale += vec3(x, y, z); }
 	/// <summary>
 	/// Scales the shape by a given amount.
 	/// </summary>
 	/// <param name="vect">Value to scale by on the 3 axes stored as a glm::vec3.</param>
-	inline void scale(glm::vec3 vect) { active_scaling += vect; }
+	inline void scale(vec3 vect) { properties.scale += vect; }
 	/// <summary>
 	/// Scales the shape by a given amount.
 	/// </summary>
 	/// <param name="val">Value to scale on all axes.</param>
-	inline void scale(float val) { active_scaling += glm::vec3(val); }
+	inline void scale(float val) { properties.scale += vec3(val); }
 #pragma endregion
 };