+++ core_gearworks_lib/source/lib/presets/2dim/rectangle/p2d_rectangle.h
@@ -20,7 +20,7 @@ public:
 	/// <param name="scale">The scale values of the rectangle primitive.</param>
 	/// <param name="texture_path">The texture to use for the rectangle.</param>
 	/// <param name="blend_colour">The colour to render the rectangle. Default is white.</param>
-	void create(gearworks::renderer *renderer, vec3 position, glm::vec3 scale, std::string texture_path, vec4 blend_colour = vec4(255, 255, 255, 255));
+	void create(gearworks::renderer *renderer, vec3 position, glm::vec3 scale, string texture_path, vec3 blend_colour = vec3(255, 255, 255), float blend_opacity = 255);
 };
 
 #endif // header guard