+++ core_gearworks_lib/source/lib/presets/2dim/triangle/p2d_triangle.h
@@ -19,7 +19,7 @@ public:
 	/// <param name="vertices">The starting vertex coordinates to use.</param>
 	/// <param name="texture_path">The texture to use for the rectangle.</param>
 	/// <param name="blend_colour">The colour to render the rectangle. Default is white.</param>
-	void create(gearworks::renderer *renderer, vec2 vertices[3], std::string texture_path, vec4 blend_colour = vec4(255, 255, 255, 255));
+	void create(gearworks::renderer *renderer, vec2 vertices[3], string texture_path, vec3 blend_colour = vec3(255, 255, 255), float blend_opacity = 255);
 };
 
 #endif // header guard